BCFN-24 fix wgrib2 calculated data
wgrib2 ensemble processing didn't appear to be calculating
correctly, especially percentile data. To fix this, statistics are
now calculated using numpy and the data is stored as a csv.

This obviously required changes to bias correction in opening old files,
but simplifies that process overall.

Percentiles are no longer limited to 10, 25, 50, 75, 90.

+++ src/ens_processing.py
@@ -66,7 +66,9 @@ def delete_old_folders():
             shutil.rmtree(folder)
     temp_files = glob(f'{DIR}tmp/*')
     for i in temp_files:
-        os.remove(i)
+        tm = dt.strptime(i.split('_')[0], '%Y%m%d%H')
+        if tm < now - timedelta(days=2):
+            os.remove(i)
 
 
 def main(args):
